% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_name_in_df.R
\name{add_name_in_df}
\alias{add_name_in_df}
\title{Agrega columnas a los dataframes de una lista basadas en su nombre}
\usage{
add_name_in_df(lista_dfs)
}
\arguments{
\item{lista_dfs}{Una lista de dataframes con nombres en el formato
`"palabra1_palabra2_numero1_numero2_numero3"`.}
}
\value{
Una lista de dataframes con las nuevas columnas agregadas.
}
\description{
Esta función toma una lista de dataframes cuyos nombres tienen el formato
`"palabra1_palabra2_numero1_numero2_numero3"`, y agrega cinco columnas nuevas
a cada dataframe: `nivel`, `indice`, `anio_inicio`, `anio_final` y `periodo`,
basadas en la estructura del nombre del dataframe.
}
\examples{
lista_dfs <- list(
  "basico_educacion_2010_2020_1" = data.frame(a = 1:3, b = 4:6),
  "medio_salud_2005_2015_2" = data.frame(x = c("A", "B"), y = c(10, 20))
)
lista_procesada <- procesar_lista_dfs(lista_dfs)
print(lista_procesada)

}
