% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_number.R
\name{calculate_number}
\alias{calculate_number}
\title{Compute a normalized row count based on unique values in selected columns}
\usage{
calculate_number(df, default_columns)
}
\arguments{
\item{df}{A dataframe containing the data.}

\item{default_columns}{A character vector specifying the columns to exclude from the calculation.
Default is `c("unidades", "datos")`.}
}
\value{
A numeric value representing the total rows divided by the product of unique values in the other columns.
        Returns `NA` if division by zero occurs.
}
\description{
This function calculates a number by dividing the total number of rows in a dataframe
by the product of the unique values in all columns except the specified default columns.
}
\examples{
df <- data.frame(unidades = rep(1:2, 4),
                 periodo = rep(1:2, each = 4),
                 sexo = rep(c("M", "F"), each = 2, times = 2),
                 turno = rep(c("M", "V"), times = 4),
                 datos = rnorm(8))
calculate_number(df)

}
