% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_acronyms.R
\name{capitalize_acronyms}
\alias{capitalize_acronyms}
\title{Capitalize Acronyms and Title Case Text}
\usage{
capitalize_acronyms(text)
}
\arguments{
\item{text}{A character string where the initial uppercase acronym will remain
capitalized, and the rest of the text will be converted to title case.}
}
\value{
A character string with the acronym in uppercase and the remaining text
        in title case.
}
\description{
This function capitalizes acronyms (uppercase sequences) and formats the remaining
text in title case. It can be useful for text formatting where acronyms need to
remain in uppercase while other parts of the text are capitalized.
}
\examples{
capitalize_acronyms('ESIA ZACATENCO')
capitalize_acronyms('CICS MILPA ALTA')
}
