% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_ID.R
\name{change_ID}
\alias{change_ID}
\title{Change IDs in Data Frame}
\usage{
change_ID(bad_data, data_ids)
}
\arguments{
\item{bad_data}{A data frame that contains the original data with the `ID_programa` column.}

\item{data_ids}{A data frame containing the new ID mappings.}
}
\value{
A data frame with the updated `ID_programa` values.
}
\description{
This function replaces the `ID_programa` column in a data frame by performing
a left join with another data frame containing ID mappings. It fills in missing
values in `ID_programa` using specified columns from the second data frame.
}
\examples{
bad_data <- data.frame(ID = c(1, 2, 3), ID_programa = c(NA, "A", NA))
data_ids <- data.frame(ID = c(1, 2), ID_programa_1 = c("X", "Y"), ID_programa_2 = c(NA, "Z"))
updated_data <- change_ID(bad_data, data_ids)
}
