% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_word_matches.R
\name{count_word_matches}
\alias{count_word_matches}
\title{Count Word Matches Between Two Columns}
\usage{
count_word_matches(df, col1, col2)
}
\arguments{
\item{df}{A data frame containing the columns to be analyzed.}

\item{col1}{A character string specifying the name of the first column.}

\item{col2}{A character string specifying the name of the second column.}
}
\value{
A numeric vector where each element represents the count of matching words
        for each row between \code{col1} and \code{col2}.
}
\description{
This function counts the number of matching words between two specified columns
in each row of a data frame.
}
\examples{
df <- data.frame(Text1 = c("apple orange", "banana pear"),
                 Text2 = c("apple banana", "pear mango"))
count_word_matches(df, "Text1", "Text2")
}
