% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_non_empty_dfs.R
\name{filter_non_empty_dfs}
\alias{filter_non_empty_dfs}
\title{Filter a list of dataframes to remove empty ones}
\usage{
filter_non_empty_dfs(df_list)
}
\arguments{
\item{df_list}{A list of dataframes.}
}
\value{
A list containing only the dataframes that have more than zero rows.
}
\description{
This function takes a list of dataframes and removes any dataframe
that has zero rows.
}
\examples{
df1 <- data.frame(a = 1:3, b = 4:6)
df2 <- data.frame()  # Empty dataframe
df3 <- data.frame(x = c("A", "B"), y = c(10, 20))

df_list <- list(df1 = df1, df2 = df2, df3 = df3)
filtered_list <- filter_non_empty_dfs(df_list)
print(filtered_list)  # df2 is removed

}
