% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest_match.R
\name{find_closest_match}
\alias{find_closest_match}
\title{Find Closest Match in Comparison Column}
\usage{
find_closest_match(search_column, compare_column)
}
\arguments{
\item{search_column}{A character vector containing the elements to be searched for
matches.}

\item{compare_column}{A character vector containing the elements to be compared
against.}
}
\value{
A data frame with three columns:
        \describe{
          \item{columna_a_buscar}{The original element from the search_column.}
          \item{coincidencia}{The closest matching element from the compare_column.}
          \item{distancia_minima}{The Levenshtein distance to the closest match.}
        }
}
\description{
This function finds the closest match for each element in a specified search column
by comparing it to another column using the Levenshtein distance.
}
\examples{
find_closest_match(c("apple", "banana"), c("apples", "banana", "berry"))
}
