% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest_match_self.R
\name{find_closest_match_self}
\alias{find_closest_match_self}
\title{Find Closest Match}
\usage{
find_closest_match_self(search_column)
}
\arguments{
\item{search_column}{A character vector containing the elements to be searched for
matches.}
}
\value{
A data frame with three columns:
        \describe{
          \item{nombre}{The original element from the search_column.}
          \item{coincidencia}{The closest matching element.}
          \item{distancia_minima}{The Levenshtein distance to the closest match.}
        }
}
\description{
This function finds the closest match for each element in a given character vector
based on the Levenshtein distance. It excludes the current element when searching for
the closest match.
}
\examples{
find_closest_match(c("apple", "apply", "apples", "banana"))
}
