% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_row.R
\name{find_row}
\alias{find_row}
\title{Find Row Containing a Specific Word}
\usage{
find_row(dataframe, word = "MIXTA", start_col = 4, end_col = 10)
}
\arguments{
\item{dataframe}{A data frame in which to search for the word.}

\item{word}{A character string specifying the word to search for. Defaults to "MIXTA".}

\item{start_col}{An integer indicating the starting column index for the search.}

\item{end_col}{An integer indicating the ending column index for the search.}
}
\value{
An integer representing the row index where the word is found, or 0 if not found.
}
\description{
This function searches for a specified word within a range of columns in a data
frame and returns the index of the first row where the word is found.
}
\examples{
df <- data.frame(A = c("a", "b", "MIXTA"), B = c("x", "MIXTA", "z"))
row_index <- find_row(df, "MIXTA", 1, 2)
}
