% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_completely_empty.R
\name{is_completely_empty}
\alias{is_completely_empty}
\title{Check if a List or Subset is Completely Empty}
\usage{
is_completely_empty(lst)
}
\arguments{
\item{lst}{A list or dataframe subset to check.}
}
\value{
A logical value: TRUE if the list or subset is completely empty, FALSE otherwise.
}
\description{
This function checks whether a given list or subset of a dataframe is completely empty.
It considers elements as empty if they are NULL or NA.
}
\examples{
empty_list <- list()
partially_empty_list <- list(NULL, 1)
completely_empty_list <- list(NULL, NULL)
dataframe_subset <- data.frame(X1 = NA, X2 = NA, X3 = NA)

is_completely_empty(empty_list)           # TRUE
is_completely_empty(partially_empty_list) # FALSE
is_completely_empty(completely_empty_list) # TRUE
is_completely_empty(dataframe_subset)     # TRUE
}
