% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posgrado.R
\name{posgrado}
\alias{posgrado}
\title{Check for Graduate Program Keywords}
\usage{
posgrado(column)
}
\arguments{
\item{column}{A character vector representing the column to check for keywords.}
}
\value{
A logical vector where each element is \code{TRUE} if the corresponding
        value in \code{column} starts with one of the specified keywords,
        indicating a graduate program, and \code{FALSE} otherwise.
}
\description{
This function checks if the values in a specified column contain keywords
indicating graduate programs, such as "Maestría," "Doctorado," or "Especialidad."
}
\examples{
column <- c("Maestría en Ciencias", "Doctorado en Filosofía", "Licenciatura en Derecho")
check_graduate_program(column)
}
