% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_datos.R
\name{procesar_datos}
\alias{procesar_datos}
\title{Procesar Datos Tabulares}
\usage{
procesar_datos(datos, columnas_creadas, n_columna)
}
\arguments{
\item{datos}{Un dataframe con los datos iniciales a procesar.}

\item{columnas_creadas}{Un vector de nombres que se utilizarán al dividir las columnas.}

\item{n_columna}{Índice de la columna clave utilizada para filtrar valores NA y reorganizar los datos.}
}
\value{
Un dataframe procesado y reorganizado.
}
\description{
Esta función procesa un dataframe eliminando columnas y filas irrelevantes,
renombrando columnas y reorganizando los datos en formato largo.
}
\examples{
datos <- data.frame(
  Col1 = c("No", "A", "B", "C"),
  Col2 = c("Unidad1", NA, "Unidad2", "Unidad3"),
  Col3 = c("Programa1", "Programa2", "Programa3", "Programa4"),
  Col4 = c(NA, 10, 20, 30)
)
columnas_creadas <- c("Concepto", "Sexo")
n_columna <- 3
datos_procesados <- procesar_datos(datos, columnas_creadas, n_columna)
}
