% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_columns.R
\name{remove_columns}
\alias{remove_columns}
\title{Remove Columns by Keyword}
\usage{
remove_columns(data, keyword)
}
\arguments{
\item{data}{A data frame from which columns will be removed.}

\item{keyword}{A character string representing the keyword to search for in column names.}
}
\value{
A data frame with the specified columns removed. If no columns match the keyword, the original data frame is returned.
}
\description{
This function removes columns from a data frame based on the presence of a specific keyword in the column names.
}
\examples{
df <- data.frame(A_key = 1:3, B = 4:6, C_key = 7:9)
remove_columns(df, "key")
}
