% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_columns_with_nulls.R
\name{remove_columns_with_nulls}
\alias{remove_columns_with_nulls}
\title{Remove Columns with Too Few Non-Null Values}
\usage{
remove_columns_with_nulls(df, max_non_nulls = 0)
}
\arguments{
\item{df}{A data frame from which columns will be removed.}

\item{max_non_nulls}{An integer specifying the minimum number of non-null values
required for a column to be retained.}
}
\value{
A data frame containing only the columns that have at least the specified
        number of non-null values.
}
\description{
This function removes columns from a data frame that have fewer than a specified
number of non-null (non-NA) values.
}
\examples{
df <- data.frame(A = c(1, NA, 3), B = c(NA, NA, NA), C = c(4, 5, 6))
filtered_df <- remove_columns_with_too_few_non_nulls(df, max_non_nulls = 2)
}
