% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_rows_below.R
\name{remove_rows_below}
\alias{remove_rows_below}
\title{Remove Rows Below a Specific Word}
\usage{
remove_rows_below(dataframe, word, column_index)
}
\arguments{
\item{dataframe}{A data frame from which rows will be removed.}

\item{word}{A character string specifying the word to search for.}

\item{column_index}{An integer representing the index of the column to search in.}
}
\value{
A data frame with all rows below the specified word removed.
}
\description{
This function removes all rows below the last occurrence of a specified word in
a given column of a data frame.
}
\examples{
df <- data.frame(A = c("start", "data", "info", "end"), B = 1:4)
cleaned_df <- remove_rows_below(df, "data", 1)
}
