% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_column.R
\name{rename_column}
\alias{rename_column}
\title{Rename Column Based on Reference Data Frame}
\usage{
rename_column(df_to_modify, reference_df)
}
\arguments{
\item{df_to_modify}{A data frame with column names to potentially modify.}

\item{reference_df}{A reference data frame with the correct column names.}
}
\value{
A data frame with the column name updated to match the reference data frame.
        If there is not exactly one differing column, the function returns
        the data frame unmodified and outputs a message.
}
\description{
This function renames a single column in a data frame based on the column names
in a reference data frame. If exactly one column name differs between the two
data frames, the differing column in the data frame to modify will be renamed
to match the name in the reference data frame.
}
\examples{
df1 <- data.frame(Name = c("A", "B"), Age = c(25, 30))
df2 <- data.frame(Nombre = c("A", "B"), Age = c(25, 30))
rename_based_on_reference(df2, df1)
}
