% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{totals}
\alias{totals}
\title{Process a dataframe by renaming, cleaning, and restructuring columns.}
\usage{
totals(data, created_columns, n_column)
}
\arguments{
\item{data}{A dataframe to process.}

\item{created_columns}{A character vector with names of new columns to create.}

\item{n_column}{A numeric value indicating a specific column index for filtering.}
}
\value{
A processed dataframe.
}
\description{
Process a dataframe by renaming, cleaning, and restructuring columns.
}
